Set the path of the JWT cookie to /
We need to set the JWT cookie's path to / to send the value of the JWT
token when making API calls targeting the deviceconnect service, which
exposes an endpoint which upgrades the connection to a websocket.

Unfortunately, browsers are not able to manipulate headers when
connecting to a websocket, thus we use the Cookie to authenticate
providing the value of the JWT user token.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>